Update best-practices-organizing.md
In line 16, 1st sentence Before creating a team, it is a good practice to think about a goal, project, or work item. and the organization’s people or groups who - s/b work item, and the organization's.. You have a period after item and it should be a comma. Thanks! Fran